add flutter feature and rename mobile to flutter
