+++ src/lib.rs
@@ -21,12 +21,12 @@ pub mod ipc;
 pub mod ui;
 mod version;
 pub use version::*;
-// #[cfg(any(target_os = "android", target_os = "ios"))]
+#[cfg(any(target_os = "android", target_os = "ios", feature = "flutter"))]
 mod bridge_generated;
-// #[cfg(any(target_os = "android", target_os = "ios"))]
-pub mod mobile;
-// #[cfg(any(target_os = "android", target_os = "ios"))]
-pub mod mobile_ffi;
+#[cfg(any(target_os = "android", target_os = "ios", feature = "flutter"))]
+pub mod flutter;
+#[cfg(any(target_os = "android", target_os = "ios", feature = "flutter"))]
+pub mod flutter_ffi;
 use common::*;
 #[cfg(feature = "cli")]
 pub mod cli;
